refactor: rename and move both platform/project usage services
