+++ packages/server/api/src/app/ee/projects/project-usage-service.ts
@@ -1,16 +1,16 @@
 import { AppSystemProp, exceptionHandler } from '@activepieces/server-shared'
 import { ApEdition, ApEnvironment, getCurrentBillingPeriodEnd, getCurrentBillingPeriodStart, isNil, ProjectUsage, UsageEntityType, UsageMetric } from '@activepieces/shared'
 import { FastifyBaseLogger } from 'fastify'
-import { getRedisConnection } from '../../../database/redis-connection'
-import { system } from '../../../helper/system/system'
-import { getUsage, redisKeyGenerator } from '../../../helper/usage'
-import { platformService } from '../../../platform/platform.service'
-import { projectService } from '../../../project/project-service'
-import { userInvitationsService } from '../../../user-invitations/user-invitation.service'
-import { platformBillingService } from '../../platform-billing/platform-billing.service'
-import { platformUsageService } from '../../platform-billing/usage/usage-service'
-import { projectMemberService } from '../../project-members/project-member.service'
-import { projectLimitsService } from '../../project-plan/project-plan.service'
+import { getRedisConnection } from '../../database/redis-connection'
+import { system } from '../../helper/system/system'
+import { getUsage, redisKeyGenerator } from '../../helper/usage'
+import { platformService } from '../../platform/platform.service'
+import { projectService } from '../../project/project-service'
+import { userInvitationsService } from '../../user-invitations/user-invitation.service'
+import { platformBillingService } from '../platform-billing/platform-billing.service'
+import { platformUsageService } from '../platform-billing/platform-usage-service'
+import { projectMemberService } from '../project-members/project-member.service'
+import { projectLimitsService } from '../project-plan/project-plan.service'
 
 const environment = system.get(AppSystemProp.ENVIRONMENT)
 const edition = system.getEdition()
