+++ packages/server/api/src/app/webhooks/webhook.service.ts
@@ -2,7 +2,7 @@ import { pinoLogging } from '@activepieces/server-shared'
 import { ActivepiecesError, apId, EngineHttpResponse, ErrorCode, EventPayload, Flow, FlowStatus, isNil, RunEnvironment, TriggerPayload, UsageMetric } from '@activepieces/shared'
 import { FastifyBaseLogger } from 'fastify'
 import { StatusCodes } from 'http-status-codes'
-import { projectUsageService } from '../ee/projects/project-usage/project-usage-service'
+import { projectUsageService } from '../ee/projects/project-usage-service'
 import { flowService } from '../flows/flow/flow.service'
 import { engineResponseWatcher } from '../workers/engine-response-watcher'
 import { handshakeHandler } from './handshake-handler'
