+++ packages/server/api/src/app/workers/engine-controller.ts
@@ -4,7 +4,7 @@ import { FastifyPluginAsyncTypebox, Type } from '@fastify/type-provider-typebox'
 import { FastifyBaseLogger } from 'fastify'
 import { StatusCodes } from 'http-status-codes'
 import { entitiesMustBeOwnedByCurrentProject } from '../authentication/authorization'
-import { projectUsageService } from '../ee/projects/project-usage/project-usage-service'
+import { projectUsageService } from '../ee/projects/project-usage-service'
 import { fileService } from '../file/file.service'
 import { flowService } from '../flows/flow/flow.service'
 import { flowRunService } from '../flows/flow-run/flow-run-service'
